Improve tests and coverage for Restorer
I mistakenly added the tests in SnapshotTest, hence here is created
a new test class RestorerTest covering half of the class Restorer.
